Organize recipes, do not load recipes in __init__.py
