+++ megaradrp/recipes/calibration/bpm.py
@@ -1,0 +1,48 @@
+#
+# Copyright 2011-2015 Universidad Complutense de Madrid
+#
+# This file is part of Megara DRP
+#
+# Megara DRP is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Megara DRP is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Megara DRP.  If not, see <http://www.gnu.org/licenses/>.
+#
+
+"""Bad PIxel Mask (BPM) recipe"""
+
+import logging
+
+from numina.core import Product
+from numina.core.requirements import ObservationResultRequirement
+
+from megaradrp.core import MegaraBaseRecipe
+from megaradrp.products import MasterBias
+
+
+_logger = logging.getLogger('numina.recipes.megara')
+
+
+class BadPixelsMaskRecipe(MegaraBaseRecipe):
+
+    '''Process BIAS images and create MASTER_BIAS.'''
+
+    obresult = ObservationResultRequirement()
+
+    biasframe = Product(MasterBias)
+
+    def __init__(self):
+        super(BadPixelsMaskRecipe, self).__init__(
+            version="0.1.0"
+        )
+
+    def run(self, rinput):
+        pass
