+++ megaradrp/recipes/scientific.py
@@ -33,8 +33,6 @@ from megaradrp.core import MegaraBaseRecipe
 from megaradrp.processing import OverscanCorrector, TrimImage
 from megaradrp.processing import ApertureExtractor, FiberFlatCorrector
 from megaradrp.processing import ApertureExtractor2
-
-# from numina.logger import log_to_history
 from megaradrp.requirements import MasterBiasRequirement
 from megaradrp.requirements import MasterFiberFlatRequirement
 from megaradrp.products import MasterFiberFlat
